refactor: Move bucketizers to their dedicated files
